feat: add option to disable CSIDriver Helm hooks (#197)
* feat: add option to disable CSIDriver Helm hooks

Add csidriver.disableHooks configuration option to allow users to disable
Helm hooks on the CSIDriver resource. This addresses issues with ArgoCD
and other GitOps tools where the combination of:
- helm.sh/hook: pre-install,pre-upgrade
- helm.sh/hook-delete-policy: before-hook-creation

causes the CSIDriver resource to be deleted and recreated on every sync,
potentially causing it to go missing if the creation fails or sync is
interrupted. This leads to PVC mount failures and cluster instability.

When csidriver.disableHooks is set to true, the CSIDriver will be
created as a regular resource without hooks, preventing the
deletion/recreation cycle.
---------

Signed-off-by: Xuhui zhang <xuhui@juicedata.io>
Co-authored-by: Zzde <xuhui@juicedata.io>